requirements-test.txt
flake8
isort
jedi
mypy
nose==1.3.4
psycopg2==2.7.5
pygments
pylint-django==0.6
pytest-benchmark==3.1.1
pytest-django==3.4.3
pytest==3.8.2
restructuredtext-lint
tox==2.9.1
